add httpattack client
