+++ src/utils/templates/templates.go
@@ -30,6 +30,10 @@ func SetProxiesURL(url string) {
 	proxiesURL = url
 }
 
+func SetProxies(list string) {
+
+}
+
 func getProxylist() (urls []string) {
 	return getProxylistByURL(getProxylistURL())
 }
